Fix service filter
