+++ packages/react-ui/src/app/features/templates/lib/tests/templates-hooks.test.ts
@@ -1,7 +1,4 @@
-import {
-  getUniqueCategoriesFromTemplates,
-  replaceServicePrefix,
-} from '../templates-hooks';
+import { getUniqueCategoriesFromTemplates } from '../templates-hooks';
 
 const makeTemplate = (partial: Partial<any>): any => ({
   name: 'dummy',
@@ -94,23 +91,5 @@ describe('getUniqueCategoriesFromTemplates', () => {
     },
   ])('$desc', ({ input, expected }) => {
     expect(getUniqueCategoriesFromTemplates(input)).toEqual(expected);
-  });
-});
-
-describe('replaceServicePrefix', () => {
-  it.each([
-    [undefined, ''],
-    ['', ''],
-    ['   ', '   '],
-    ['AWS EC2', 'EC2'],
-    ['GCP BigQuery', 'BigQuery'],
-    ['Azure Cosmos DB', 'Cosmos DB'],
-    ['AWS   Lambda', 'Lambda'],
-    ['AWS', 'AWS'],
-    ['DigitalOcean Spaces', 'DigitalOcean Spaces'],
-    ['EC2', 'EC2'],
-    ['AWS-EC2', 'AWS-EC2'],
-  ])('replaces "%s" with "%s"', (input, expected) => {
-    expect(replaceServicePrefix(input)).toBe(expected);
   });
 });
